smartbots-etl/pyproject.toml
[project]
name = "smartbots-etl"
version = "0.1.0"
description = "ETL de consolidación de facturas XLSX via Google Drive"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.2",
    "openpyxl>=3.1",
    "google-api-python-client>=2.160",
    "google-auth>=2.37",
    "structlog>=24.4",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-cov>=6.0",
    "ruff>=0.9",
    "mypy>=1.14",
    "pandas-stubs>=2.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "UP", "S", "B", "A", "C4", "RUF"]
ignore = ["S101"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--strict-markers --tb=short -q"
markers = ["integration: tests que requieren I/O real"]

[tool.mypy]
python_version = "3.12"
strict = true
